+++ src/main/java/ddingdong/ddingdongBE/domain/club/controller/dto/request/UpdateClubMemberRequest.java
@@ -1,7 +1,7 @@
 package ddingdong.ddingdongBE.domain.club.controller.dto.request;
 
 import ddingdong.ddingdongBE.domain.club.entity.Position;
-import ddingdong.ddingdongBE.domain.club.service.dto.UpdateClubMemberCommand;
+import ddingdong.ddingdongBE.domain.clubmember.service.dto.UpdateClubMemberCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
