+++ src/main/java/ddingdong/ddingdongBE/domain/clubmember/service/dto/UpdateClubMemberCommand.java
@@ -1,6 +1,6 @@
-package ddingdong.ddingdongBE.domain.club.service.dto;
+package ddingdong.ddingdongBE.domain.clubmember.service.dto;
 
-import ddingdong.ddingdongBE.domain.club.entity.ClubMember;
+import ddingdong.ddingdongBE.domain.clubmember.entity.ClubMember;
 import ddingdong.ddingdongBE.domain.club.entity.Position;
 import lombok.Builder;
