+++ src/main/java/ddingdong/ddingdongBE/file/service/ExcelFileService.java
@@ -3,7 +3,7 @@ package ddingdong.ddingdongBE.file.service;
 import ddingdong.ddingdongBE.common.exception.ParsingExcelFileException.ExcelIO;
 import ddingdong.ddingdongBE.common.exception.ParsingExcelFileException.NonExcelFile;
 import ddingdong.ddingdongBE.domain.club.entity.Club;
-import ddingdong.ddingdongBE.domain.club.entity.ClubMember;
+import ddingdong.ddingdongBE.domain.clubmember.entity.ClubMember;
 import ddingdong.ddingdongBE.file.service.dto.ExcelClubMemberDto;
 import java.awt.Color;
 import java.io.ByteArrayOutputStream;
